+++ x/oracle/keeper/keeper_suite_test.go
@@ -7,6 +7,7 @@ import (
 	math "cosmossdk.io/math"
 	"github.com/ExocoreNetwork/exocore/testutil"
 	"github.com/ExocoreNetwork/exocore/x/oracle/keeper"
+	"github.com/ExocoreNetwork/exocore/x/oracle/keeper/testdata"
 	"github.com/ExocoreNetwork/exocore/x/oracle/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/suite"
@@ -72,7 +73,7 @@ func TestKeeper(t *testing.T) {
 }
 
 func (suite *KeeperSuite) Reset() {
-	p4Test := DefaultParamsForTest()
+	p4Test := testdata.DefaultParamsForTest()
 	p4Test.TokenFeeders[1].StartBaseBlock = 1
 	suite.k.SetParams(suite.ctx, p4Test)
 	suite.k.FeederManager.SetNilCaches()
@@ -105,7 +106,7 @@ func (suite *KeeperSuite) SetupTest() {
 	suite.ms = keeper.NewMsgServerImpl(suite.App.OracleKeeper)
 	suite.ctx = suite.Ctx
 	// Initialize params
-	p4Test := DefaultParamsForTest()
+	p4Test := testdata.DefaultParamsForTest()
 	p4Test.TokenFeeders[1].StartBaseBlock = 1
 	suite.k.SetParams(suite.ctx, p4Test)
 	suite.ctx = suite.ctx.WithBlockHeight(12)
